I've posted comments about lists
Do I need to put the name of process in a list? What Ima gonna do with this? 
Create a loop function; it will take the quantum process and "process" (i will make quantum = quantum - max_quantum, where max_quantum is for the bound type)
Time is ticking, time is running...